requirements.txt
pandas
numpy
apache-airflow
transformers
streamlit
requests
bs4
tqdm
requests
psycopg2